Changed the way oil generates.
Oil will generate between stone layers now, from a small pool to a huge
one that can be found on the bottom. The small, first layer, pools have
a chance of generating oil tunnels that lead to the surface. Also
changed a few recipes that had a need for the crafting table.